Move page styles to app dir
