+++ app/environment/style.module.scss
@@ -1,4 +1,4 @@
-@use 'mixins';
+@use '../../styles/mixins';
 
 .name {
     font-weight: bold;
