+++ app/music/page.mdx
@@ -5,7 +5,7 @@ import Block from "../../components/Block";
 import LiteYouTubeEmbedWrapper from "../../components/utils/LiteYouTubeEmbedWrapper";
 import Lyrics from "./Lyrics.mdx";
 
-import styles from '../../styles/music.module.scss';
+import styles from './style.module.scss';
 
 <Title title={'つまみのうた'}>
   ねぎ一世(<a href="https://twitter.com/negiissei">@negiissei</a>)さんに「<b>つまみのうた</b>」を作っていただきました！(？？？？)
