+++ app/stickers/style.module.scss
